Includes cleaning

+++ Jeu/Src/Area/AreaMap.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 #include <cmath>
 
-#include "../Game.h"
+#include "../Common/Globalbilboulga.h"
 
 const int roomMinSize = 16;
 const int roomScale = 4;
@@ -38,14 +38,14 @@ void AreaMap::drawRoom(SDL_Rect roomDim, std::vector<bool> exits)
 	{
 		if (exits[i] == 1)
 		{
-			firstExit = i;
+			firstExit = static_cast<int>(i);
 			break;
 		}
 	}
 
 	for (size_t i = 0; i < exits.size(); i++)
 	{
-		if (exits[i] == 1) lastExit = i;
+		if (exits[i] == 1) lastExit = static_cast<int>(i);
 	}
